chore(frontend): refactor members view into scoped components
