+++ apps/frontend/src/stores/members.ts
@@ -10,15 +10,10 @@ import type {
 import { addMember, getMembers } from "../services/members";
 import { useUserStore } from "../stores/user";
 
-interface PaginationState {
-  currentPage: number;
-  pageSize: number;
-  totalItems: number;
-  totalPages: number;
-  offset: number;
-  loading: boolean;
-}
-
+/**
+ * Members store focused on data management following Single Responsibility Principle
+ * Pagination UI logic is handled by parent components
+ */
 export const useMembersStore = defineStore("members", () => {
   const defaultPageLimit = 5;
   const userStore = useUserStore();
@@ -27,31 +22,19 @@ export const useMembersStore = defineStore("members", () => {
     data: [],
     totalCount: 0,
   });
-
-  // Pagination state
-  const currentPage = ref<number>(1);
-  const pageSize = ref<number>(defaultPageLimit);
   const sortOrder = ref<"asc" | "desc">("asc");
 
-  // Computed pagination values
+  // Computed values for data access
   const totalItems = computed(() => members.value.totalCount);
-  const totalPages = computed(() => Math.ceil(totalItems.value / pageSize.value));
-  const offset = computed(() => (currentPage.value - 1) * pageSize.value);
 
-  // Pagination state object
-  const paginationState = computed<PaginationState>(() => ({
-    currentPage: currentPage.value,
-    pageSize: pageSize.value,
-    totalItems: totalItems.value,
-    totalPages: totalPages.value,
-    offset: offset.value,
-    loading: loading.value,
-  }));
-
+  /**
+   * Fetch members with pagination parameters
+   * @param params - Pagination and filtering parameters
+   */
   async function getMembersPaginated(params?: IGetMembersPaginatedReq) {
     const requestParams = {
-      limit: params?.limit || `${pageSize.value}`,
-      offset: params?.offset || `${offset.value}`,
+      limit: params?.limit || `${defaultPageLimit}`,
+      offset: params?.offset || "0",
       order: params?.order || sortOrder.value,
     };
 
@@ -72,46 +55,28 @@ export const useMembersStore = defineStore("members", () => {
     loading.value = false;
   }
 
-  // Pagination helper methods
-  async function goToPage(page: number) {
-    if (page < 1 || page > totalPages.value) return;
-    currentPage.value = page;
-    await getMembersPaginated();
-  }
-
-  async function goToNextPage() {
-    if (currentPage.value < totalPages.value) {
-      currentPage.value += 1;
-      await getMembersPaginated();
-    }
-  }
-
-  async function goToPreviousPage() {
-    if (currentPage.value > 1) {
-      currentPage.value -= 1;
-      await getMembersPaginated();
-    }
-  }
-
-  async function changePageSize(newPageSize: number) {
-    pageSize.value = newPageSize;
-    currentPage.value = 1; // Reset to first page when changing page size
-    await getMembersPaginated();
-  }
-
-  async function changeSortOrder(order: "asc" | "desc") {
+  /**
+   * Change sort order and refetch data
+   * @param order - Sort order (asc or desc)
+   * @param limit - Page size
+   * @param offset - Page offset
+   */
+  async function changeSortOrder(order: "asc" | "desc", limit?: number, offset?: number) {
     sortOrder.value = order;
-    await getMembersPaginated();
+    await getMembersPaginated({
+      order,
+      limit: limit?.toString() || `${defaultPageLimit}`,
+      offset: offset?.toString() || "0",
+    });
   }
 
-  function resetPagination() {
-    currentPage.value = 1;
-    pageSize.value = defaultPageLimit;
-    sortOrder.value = "asc";
-  }
-
-  async function addNewMember(params: ICreateMemberReq) {
-    const [data, error] = await addMember(params, userStore.accessToken);
+  /**
+   * Add a new member and refresh the current page
+   * @param memberData - Member data to create
+   * @param currentParams - Current pagination parameters for refresh
+   */
+  async function addNewMember(memberData: ICreateMemberReq, currentParams?: IGetMembersPaginatedReq) {
+    const [data, error] = await addMember(memberData, userStore.accessToken);
 
     if (error) {
       toast.error("Unexpected error while adding a member");
@@ -121,27 +86,19 @@ export const useMembersStore = defineStore("members", () => {
     toast(
       `New member added: ${data?.firstName} ${data?.lastName}`,
     );
-    await getMembersPaginated();
+
+    // Refresh with current pagination state
+    await getMembersPaginated(currentParams);
   }
 
   return {
     members,
     loading,
     defaultPageLimit,
-    paginationState,
-    currentPage,
-    pageSize,
     sortOrder,
     totalItems,
-    totalPages,
-    offset,
     getMembersPaginated,
     addNewMember,
-    goToPage,
-    goToNextPage,
-    goToPreviousPage,
-    changePageSize,
     changeSortOrder,
-    resetPagination,
   };
 });
